third_party/WebKit/LayoutTests/virtual/feature-policy-vibrate/http/tests/feature-policy-vibrate/vibrate-enabledforself-expected.txt

CONSOLE ERROR: line 15: Navigator.vibrate() is not enabled in feature policy for this frame.
This is a testharness.js-based test.
PASS Navigator.vibrate enabled for self on URL: resources/feature-policy-vibrate.html
PASS Navigator.vibrate enabled for self on URL: http://localhost:8000/feature-policy-vibrate/resources/feature-policy-vibrate.html
Harness: the test ran to completion.
